backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.12.0
requests==2.31.0
anthropic==0.34.1
python-dotenv==1.0.1